Sec7/partners animation (#8)
* Implement PartnersAnimation component in React.

Created the PartnersAnimation.jsx component to transition the respective section from plain JavaScript to React. Also, added and imported the partners-animation.css file for styling.

* style: Refine styles in partners-animation.css

Made adjustments in partners-animation.css to enhance and fine-tune the styles for a better appearance.

* style: Remove comments and apply ThemeChange

Modified files to remove unnecessary comments and applied ThemeChange for consistency and improved styling.

* style: Adjust styles for responsive carousel in PartnersAnimation

Modified css and jsx files in the PartnersAnimation component to fine-tune styles, ensuring a responsive layout for the carousel without errors. Also fixed right-margin issues caused by ::after elements in the CSS.

* style: Refine styles in PartnersAnimation.jsx

Made style adjustments in the PartnersAnimation.jsx file to enhance the overall appearance of the component.

* style: Adjust styles in modified files

Made style adjustments in the modified files to enhance the overall appearance.